refactor: 구간 조회 API 기능 담당 클래스 이동
